Вынес логику, решил задачи, поправил поиск. Не нашел бага с
